LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/css/selectors/user-valid-expected.txt


Test form interactions (reset / submit):







PASS :user-valid selector should respond to user action
PASS :user-valid selector properly interacts with submit & reset buttons
PASS Checkboxes should match :user-valid after the user clicks on it.
FAIL Date inputs should match :user-valid after the user types a value into it. assert_equals: Date input value should match the testdriver input. expected "1234-01-01" but got "1"
